AI-472 Address Google ADK metrics review
